Filter logging for tx submission stages
